requirements.txt
PySide6>=6.8
openai>=1.108.0
pypdf>=5.3.0
pydantic>=2.10.0
python-dotenv>=1.0.1
